requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
Django==3.1.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
regex==2020.10.15
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
